Update index.php
If you write if ($result = AutoUpdate::ERROR_SIMULATE), then assign the value of AutoUpdate::ERROR_SIMULATE to the variable $result and then check if this value is true. This is probably not what you want. If you write if ($result == AutoUpdate::ERROR_SIMULATE), then compare the value of $result with the value of AutoUpdate::ERROR_SIMULATE and then check if these values are the same. This is probably what you want.